refactor(episode): define reusable episode atoms to prepare for up-next
